mas-api/requirements.txt
fastapi==0.95.2
gunicorn==20.1.0
uvicorn==0.22.0
SQLAlchemy==2.0.14
aiomysql==0.1.1
greenlet==2.0.2
cryptography==40.0.2
Inject==4.3.1
pytz==2023.3
google-auth==2.18.1
pydantic==1.10.7
python-jose[cryptography]==3.3.0
requests==2.30.0
httpx-oauth==0.11.2
loguru==0.7.0
